+++ typing-tests/tests/error-missing-response-header.ts
@@ -8,4 +8,4 @@ export const handler: RouteHandler<Response.Ok<
 })
 
 // Expected error:
-// Type 'RouteHandler<Context, Response<200, string, undefined>>' is not assignable to type 'RouteHandler<Context, Response<200, string, { 'Content-Type': "application/json"; }>>'.
+// Type 'RouteHandler<KoaRequestBase, Response<200, string, undefined>>' is not assignable to type 'RouteHandler<KoaRequestBase, Response<200, string, { 'Content-Type': "application/json"; }>>'.